[REF] .gitignore: remove vendor folder from tracking
