+++ php-mvc/src/controllers/post/Post.php
@@ -2,13 +2,9 @@
 
 namespace Application\Controllers\Post;
 
-require_once('src/lib/database.php');
-require_once('src/model/comment.php');
-require_once('src/model/post.php');
-
-use Application\Lib\Database\DatabaseConnection;
-use Application\Model\Comment\CommentRepository;
-use Application\Model\Post\PostRepository;
+use Application\Lib\DatabaseConnection\DatabaseConnection;
+use Application\Model\CommentRepository\CommentRepository;
+use Application\Model\PostRepository\PostRepository;
 
 class Post
 {
@@ -24,6 +20,6 @@ class Post
         $commentRepository->connection = $connection;
         $comments = $commentRepository->getComments($identifier);
 
-        require('templates/post.php');
+        require('templates/Post.php');
     }
 }
