+++ php-mvc/src/model/comment/Comment.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Application\Model\Comment;
+
+use Application\Lib\Database\DatabaseConnection;
+
+class Comment
+{
+    public string $identifier;
+    public string $author;
+    public string $frenchCreationDate;
+    public string $comment;
+    public string $post;
+}
