+++ php-mvc/src/model/post/Post.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Application\Model\Post;
+
+
+use Application\Lib\DatabaseConnection\DatabaseConnection;
+
+class Post
+{
+    public string $title;
+    public string $frenchCreationDate;
+    public string $content;
+    public string $identifier;
+}
